Make for loops over empty lists succeed
for-loops that are never executed carry over the last status.
This can be surprising in edge cases, and it's inconsistent with
both POSIX semantics and our while loops.

Fix this by special-casing empty lists.  If the for loop runs at least
once, we use the exit status from the body. Note that the body may
be empty, in which case we still carry over the status from before
the loop.

The case where two empties collide is still a bit weird:

	false
	for x in
	end
	# status is 0, due to the fix

whereas

	false
	for x in 1 2 3
	end
	# status is 1

Closes #8409