Collapse universal scope into global scope
Prior to this change, there was a separate "universal" scope above the global
scope. It was possible to have a universal variable which was shadowed by a
global of the same name.

This changes "universal" to be a *variable property*, like "exports." Only
globals may have this property set. "Universal variable" now means "global
variable that is also marked as universal." `set --universal` now implies the
`--global` option.  When you create, modify, or erase a universal global, all
other sessions change their corresponding global as well.

The most important implication is that *universal variables now take precedence
over inherited environment variables*. This addresses #5258 and #5348. In
principle this allows variables like PATH to be universal; however that's not
recommended (yet).

For compatibility, we retain `set --query --universal` which now means "is
there a global variable with the universal property."

One important implication is that `set -qg` will now succeed for universal
variables:

    > set --universal  myvar someval
    > set --query --global myvar # succeeds

Functions which expected this to fail (`abbr` is an example) need to be
updated.

Fixes #5348. Fixes #5258. Fixes #7379. Fixes #7317.