README: Link to Unix & Linux Stackexchange instead of SO
Stackoverflow's fish tag suffers from inconsistent moderation and an
annoying policy on what is allowed and what isn't.

Given that fish straddles the line between "programming" and "usage",
some fish questions would be allowed and some wouldn't, and it is
awkward for users to tell which.

So stop recommending a site that, in practice, closes user's questions
for unclear reasons.

This needs to be done for fish-site as well.