Add a flag to disable the missing-interpreter hack
It's normally expected that script files should have their interpreter
specified at the top of the file with a line like

  #!/bin/sh

This convention was introduced in 40+ years ago (in 1980), and the
kernel wisely rejects any script which doesn't have an interpreter.  The
POSIX shell retains compatibility with scripts written before then by
using a heuristic to determine if the file being executed "looks like a
shell script" and assumes the shell as an interpreter if so.

fish implements this heuristic.

Unfortunately, the heuristic in question will happily accept pretty much
anything that looks like any script at all, regardless of the language,
leading to confusing errors when (for example) trying to run a Python
script through /bin/sh.

Instead of trying to "improve" the heuristic, add an option to disable
it entirely, for people who don't mind losing compatibility with shell
scripts that were written before they were born.

Fixes #8086